+++ packages/entity_store/lib/src/storage_repository/options.dart
@@ -1,67 +1,43 @@
 part of '../storage_repository.dart';
 
-class StorageFindByIdOptions
-    implements FindByIdOptions, FetchPolicyOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
+class StorageFindByIdOptions implements FindByIdOptions, FetchPolicyOptions {
   @override
   FetchPolicy fetchPolicy;
 
   StorageFindByIdOptions({
-    this.enableBefore = true,
     this.fetchPolicy = FetchPolicy.persistent,
   });
 }
 
-class StorageFindOneOptions
-    implements FindOneOptions, FetchPolicyOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
+class StorageFindOneOptions implements FindOneOptions, FetchPolicyOptions {
   @override
   FetchPolicy fetchPolicy;
 
   StorageFindOneOptions({
-    this.enableBefore = true,
     this.fetchPolicy = FetchPolicy.persistent,
   });
 }
 
-class StorageFindAllOptions
-    implements FindAllOptions, FetchPolicyOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
+class StorageFindAllOptions implements FindAllOptions, FetchPolicyOptions {
   @override
   FetchPolicy fetchPolicy;
 
   StorageFindAllOptions({
-    this.enableBefore = true,
     this.fetchPolicy = FetchPolicy.persistent,
   });
 }
 
-class StorageUpsertOptions
-    implements UpsertOptions, FetchPolicyOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
+class StorageUpsertOptions implements UpsertOptions, FetchPolicyOptions {
   @override
   FetchPolicy fetchPolicy;
 
   StorageUpsertOptions({
-    this.enableBefore = true,
     this.fetchPolicy = FetchPolicy.persistent,
   });
 }
 
-class StorageSaveOptions
-    implements SaveOptions, BeforeCallbackOptions, LoadEntityCallbackOptions {
-  @override
+class StorageSaveOptions implements SaveOptions {
   bool enableBefore;
-
-  @override
   bool enableLoadEntity;
 
   StorageSaveOptions({
@@ -70,30 +46,14 @@ class StorageSaveOptions
   });
 }
 
-class StorageDeleteOptions implements DeleteOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
-  StorageDeleteOptions({
-    this.enableBefore = true,
-  });
+class StorageDeleteOptions implements DeleteOptions {
+  StorageDeleteOptions();
 }
 
-class StorageQueryUpsertOptions
-    implements QueryUpsertOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
-  StorageQueryUpsertOptions({
-    this.enableBefore = true,
-  });
+class StorageQueryUpsertOptions implements QueryUpsertOptions {
+  StorageQueryUpsertOptions();
 }
 
-class StorageCountOptions implements CountOptions, BeforeCallbackOptions {
-  @override
-  bool enableBefore;
-
-  StorageCountOptions({
-    this.enableBefore = true,
-  });
+class StorageCountOptions implements CountOptions {
+  StorageCountOptions();
 }