+++ packages/entity_store_sembast/lib/src/storage_repository.dart
@@ -1,5 +1,2 @@
-import 'package:entity_store/entity_store.dart';
-import 'package:sembast/sembast.dart';
-import 'package:type_result/type_result.dart';
-
 export 'repository.dart';
+export 'query.dart';